refactor: consolidate solver package into diff
