+++ cmd/utils.go
@@ -3,11 +3,11 @@ package cmd
 import (
 	"github.com/fatih/color"
 	"github.com/kong/deck/cprint"
-	"github.com/kong/deck/solver"
+	"github.com/kong/deck/diff"
 	"github.com/spf13/pflag"
 )
 
-func printStats(stats solver.Stats) {
+func printStats(stats diff.Stats) {
 	// do not use github.com/kong/deck/print because that package
 	// is used only for events logs
 	printFn := color.New(color.FgGreen, color.Bold).PrintfFunc()
