+++ diff/diff_helpers_test.go
@@ -1,4 +1,4 @@
-package solver
+package diff
 
 import (
 	"testing"
